Renamed core/queue to core/queue_config because of a name conflict with Python's built-in queue module
