+++ config/queue_registry.py
@@ -1,6 +1,6 @@
 # config/queue_registry.py
 
-from core.queue import QueueConfig
+from core.queue_config import QueueConfig
 
 # Users can modify this file to register any number of queues with custom config
 # This is business-logic agnostic and suitable for open-source usage
@@ -8,6 +8,6 @@ from core.queue import QueueConfig
 REGISTERED_QUEUES = [
     QueueConfig(name="default"),
     # Example:
-    # QueueConfig(name="image_processing", priorities=["high", "low"], retry_limit=5),
+    QueueConfig(name="image_processing", priorities=["high", "low"]),
     # QueueConfig(name="billing", enable_dlq=False),
 ]
